Quotes aren't grids
OK, so the only way I can get the quotes to keep a tight border for a speech bubble is if they are not inside a grid, which makes everything else terrible. But hey, maybe we can do some media query magic to make it work. Either way, this is a known decent state for 1200px screens and OK-ish below that.